+++ tests/testthat/test_uci.R
@@ -6,10 +6,12 @@ grid <- readRDS(file.path(data_dir, "grid_bho.rds"))
 
 tester <- function(sf_object = grid,
                    var_name = 'jobs',
+                   dist_type = 'euclidean',
                    bootstrap_border = FALSE,
                    showProgress = TRUE) {
   uci(sf_object,
       var_name,
+      dist_type,
       bootstrap_border,
       showProgress)
 }
@@ -21,6 +23,9 @@ test_that("expected behavior", {
   result <- tester()
   testthat::expect_is(result, "data.frame")
 
+  result <- tester(dist_type = 'spatial_link')
+  testthat::expect_is(result, "data.frame")
+  
   result <- tester(bootstrap_border = TRUE, showProgress = TRUE)
   testthat::expect_is(result, "data.frame")
   
@@ -36,6 +41,8 @@ test_that("raises errors due to incorrect input", {
 
   testthat::expect_error(tester(sf_object='banana'))
   testthat::expect_error(tester(var_name = 'banana'))
+  testthat::expect_error(tester(dist_type = 'banana'))
+  testthat::expect_error(tester(dist_type = 123))
   testthat::expect_error(tester(bootstrap_border = 'banana'))
   testthat::expect_error(tester(showProgress = 'banana'))